+++ CHANGES.md
@@ -38,6 +38,8 @@
 
   + Fix comments around desugared expression (#1487, @gpetiot)
 
+  + Fix `ocamlformat.el` buffer replacement for MacOS Emacs (#1481, @juxd)
+
 #### New features
 
 ### 0.15.0 (2020-08-06)